// sdks/typescript/src/apache_beam/proto/schema.ts
export interface Schema {
    /**
     * List of fields for this schema. Two fields may not share a name.
     *
     * @generated from protobuf field: repeated org.apache.beam.model.pipeline.v1.Field fields = 1;
     */
    fields: Field[];
    /**
     * REQUIRED. An RFC 4122 UUID.
     *
     * @generated from protobuf field: string id = 2;
     */
    id: string;
    /**
     * @generated from protobuf field: repeated org.apache.beam.model.pipeline.v1.Option options = 3;
     */
    options: Option[];
    /**
     * Indicates that encoding positions have been overridden.
     *
     * @generated from protobuf field: bool encoding_positions_set = 4;
     */
    encodingPositionsSet: boolean;
}
/**
 * @generated from protobuf message org.apache.beam.model.pipeline.v1.Field
 */
export interface Field {
    /**
     * REQUIRED. Name of this field within the schema.
     *
     * @generated from protobuf field: string name = 1;
     */
    name: string;
    /**
     * OPTIONAL. Human readable description of this field, such as the query that generated it.
     *
     * @generated from protobuf field: string description = 2;
     */
    description: string;
    /**
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.FieldType type = 3;
     */
    type?: FieldType;
    /**
     * @generated from protobuf field: int32 id = 4;
     */
    id: number;
    /**
     * OPTIONAL. The position of this field's data when encoded, e.g. with beam:coder:row:v1.
     * Either no fields in a given row are have encoding position populated,
     * or all of them are. Used to support backwards compatibility with schema
     * changes.
     * If no fields have encoding position populated the order of encoding is the same as the order in the Schema.
     * If this Field is part of a Schema where encoding_positions_set is True then encoding_position must be
     * defined, otherwise this field is ignored.
     *
     * @generated from protobuf field: int32 encoding_position = 5;
     */
    encodingPosition: number;
    /**
     * @generated from protobuf field: repeated org.apache.beam.model.pipeline.v1.Option options = 6;
     */
    options: Option[];
}
/**
 * @generated from protobuf message org.apache.beam.model.pipeline.v1.FieldType
 */
export interface FieldType {
    /**
     * @generated from protobuf field: bool nullable = 1;
     */
    nullable: boolean;
    /**
     * @generated from protobuf oneof: type_info
     */
    typeInfo: {
        oneofKind: "atomicType";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.AtomicType atomic_type = 2;
         */
        atomicType: AtomicType;
    } | {
        oneofKind: "arrayType";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ArrayType array_type = 3;
         */
        arrayType: ArrayType;
    } | {
        oneofKind: "iterableType";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.IterableType iterable_type = 4;
         */
        iterableType: IterableType;
    } | {
        oneofKind: "mapType";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.MapType map_type = 5;
         */
        mapType: MapType;
    } | {
        oneofKind: "rowType";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.RowType row_type = 6;
         */
        rowType: RowType;
    } | {
        oneofKind: "logicalType";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.LogicalType logical_type = 7;
         */
        logicalType: LogicalType;
    } | {
        oneofKind: undefined;
    };
}
// ...
export interface Option {
    /**
     * REQUIRED. Identifier for the option.
     *
     * @generated from protobuf field: string name = 1;
     */
    name: string;
    /**
     * REQUIRED. Type specifer for the structure of value.
     * Conventionally, options that don't require additional configuration should
     * use a boolean type, with the value set to true.
     *
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.FieldType type = 2;
     */
    type?: FieldType;
    /**
     * @generated from protobuf field: org.apache.beam.model.pipeline.v1.FieldValue value = 3;
     */
    value?: FieldValue;
}
/**
 * @generated from protobuf message org.apache.beam.model.pipeline.v1.Row
 */
export interface Row {
    /**
     * @generated from protobuf field: repeated org.apache.beam.model.pipeline.v1.FieldValue values = 1;
     */
    values: FieldValue[];
}
/**
 * @generated from protobuf message org.apache.beam.model.pipeline.v1.FieldValue
 */
export interface FieldValue {
    /**
     * @generated from protobuf oneof: field_value
     */
    fieldValue: {
        oneofKind: "atomicValue";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.AtomicTypeValue atomic_value = 1;
         */
        atomicValue: AtomicTypeValue;
    } | {
        oneofKind: "arrayValue";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.ArrayTypeValue array_value = 2;
         */
        arrayValue: ArrayTypeValue;
    } | {
        oneofKind: "iterableValue";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.IterableTypeValue iterable_value = 3;
         */
        iterableValue: IterableTypeValue;
    } | {
        oneofKind: "mapValue";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.MapTypeValue map_value = 4;
         */
        mapValue: MapTypeValue;
    } | {
        oneofKind: "rowValue";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.Row row_value = 5;
         */
        rowValue: Row;
    } | {
        oneofKind: "logicalTypeValue";
        /**
         * @generated from protobuf field: org.apache.beam.model.pipeline.v1.LogicalTypeValue logical_type_value = 6;
         */
        logicalTypeValue: LogicalTypeValue;
    } | {
        oneofKind: undefined;
    };
}
// ...
export enum AtomicType {
    /**
     * @generated from protobuf enum value: UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: BYTE = 1;
     */
    BYTE = 1,
    /**
     * @generated from protobuf enum value: INT16 = 2;
     */
    INT16 = 2,
    /**
     * @generated from protobuf enum value: INT32 = 3;
     */
    INT32 = 3,
    /**
     * @generated from protobuf enum value: INT64 = 4;
     */
    INT64 = 4,
    /**
     * @generated from protobuf enum value: FLOAT = 5;
     */
    FLOAT = 5,
    /**
     * @generated from protobuf enum value: DOUBLE = 6;
     */
    DOUBLE = 6,
    /**
     * @generated from protobuf enum value: STRING = 7;
     */
    STRING = 7,
    /**
     * @generated from protobuf enum value: BOOLEAN = 8;
     */
    BOOLEAN = 8,
    /**
     * @generated from protobuf enum value: BYTES = 9;
     */
    BYTES = 9
}
// @generated message type with reflection information, may provide speed optimized methods
class Schema$Type extends MessageType<Schema> {
    constructor() {
        super("org.apache.beam.model.pipeline.v1.Schema", [
            { no: 1, name: "fields", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => Field },
            { no: 2, name: "id", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "options", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => Option },
            { no: 4, name: "encoding_positions_set", kind: "scalar", T: 8 /*ScalarType.BOOL*/ }
        ]);
    }
    create(value?: PartialMessage<Schema>): Schema {
        const message = { fields: [], id: "", options: [], encodingPositionsSet: false };
        globalThis.Object.defineProperty(message, MESSAGE_TYPE, { enumerable: false, value: this });
        if (value !== undefined)
            reflectionMergePartial<Schema>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Schema): Schema {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated org.apache.beam.model.pipeline.v1.Field fields */ 1:
                    message.fields.push(Field.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* string id */ 2:
                    message.id = reader.string();
                    break;
                case /* repeated org.apache.beam.model.pipeline.v1.Option options */ 3:
                    message.options.push(Option.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* bool encoding_positions_set */ 4:
                    message.encodingPositionsSet = reader.bool();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Schema, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated org.apache.beam.model.pipeline.v1.Field fields = 1; */
        for (let i = 0; i < message.fields.length; i++)
            Field.internalBinaryWrite(message.fields[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* string id = 2; */
        if (message.id !== "")
            writer.tag(2, WireType.LengthDelimited).string(message.id);
        /* repeated org.apache.beam.model.pipeline.v1.Option options = 3; */
        for (let i = 0; i < message.options.length; i++)
            Option.internalBinaryWrite(message.options[i], writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        /* bool encoding_positions_set = 4; */
        if (message.encodingPositionsSet !== false)
            writer.tag(4, WireType.Varint).bool(message.encodingPositionsSet);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message org.apache.beam.model.pipeline.v1.Schema
 */
export const Schema = new Schema$Type();
// ...
export const Field = new Field$Type();
// ...
export const FieldType = new FieldType$Type();
// ...
export const Option = new Option$Type();
// ...
export const FieldValue = new FieldValue$Type();